Compile v8env with tsc, restructure a bit
