+++ v8env/src/blob.ts
