+++ v8env/src/console.js
