+++ v8env/src/fly/static.ts
