+++ v8env/src/index.js
@@ -12,10 +12,10 @@ import Headers from './headers'
 
 import { TextEncoder, TextDecoder } from './text-encoding'
 import { fetch } from './fetch'
-import Body from './ts/body_mixin.ts'
-import Blob from './ts/blob.ts'
-import FormData from './ts/form_data.ts'
-import { crypto } from './ts/crypto.ts'
+import Body from './body_mixin'
+import Blob from './blob'
+import FormData from './form_data'
+import { crypto } from './crypto'
 import { Response } from './response'
 import { Request } from './request'
 import cache from './cache'
