+++ v8env/src/streams.js
